refactor(web): extract wire-types and drop residual passkey cast (review)
Extract shared Better Auth wire shapes (`AuthApiResult`, `Passkey`,
`PasskeySignIn`, `PasskeySignInData`) into `lib/auth/wire-types.ts`
so the boundary type in `client.ts` and the per-plugin helper
(`passkey-client.ts`) reference identical types. Without this the
boundary returned an inline `Promise<AuthResult<{session, user}>>`
and the helper-local `PasskeySignIn` returned the same shape under
a different alias — forcing `as PasskeySignIn` at the helper return.

Also:
- Drop three near-identical "typed at the export boundary" comments
  at consumer sites (kept the rationale in `client.ts` only).
- Trim the passkey-client.ts file docstring to the parts that aren't
  obvious from the code.
