+++ packages/web/src/lib/auth/two-factor-client.ts
@@ -56,9 +56,6 @@ export interface TwoFactorClient {
  * catches Better Auth API drift at the boundary.
  */
 export function getTwoFactorClient(): TwoFactorClient | null {
-  // `authClient.twoFactor` is typed at the export boundary in
-  // `lib/auth/client.ts`; the runtime guard still catches Better Auth
-  // method renames at the boundary.
   const namespace = authClient.twoFactor;
   if (
     !namespace ||
